requirements.txt
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
alembic==1.13.0
asyncpg==0.27.0
case-converter==1.1.0
celery==5.2.7
celery-sqlalchemy-scheduler==0.3.0
email-validator==2.0
fastapi==0.109.1
fastapi-async-sqlalchemy==0.6.0
fastapi-limiter==0.1.5
fastapi-nextauth-jwt==1.1.2
fastapi-pagination[sqlalchemy]==0.12.14
httpx==0.23.1
isort==5.13.2
langchain==0.1.0
langchain-community==0.0.11
langchain-openai==0.0.2
langsmith==0.0.78
markdown==3.6
minio==7.1.13
networkx==3.3
numpy==1.0.0
openai==1.6.1
opencv-python==4.10.0.84
openpyxl==3.0.10
pandas==2.2.2
passlib==1.7.4
pgvector==0.1.6
pillow==10.4.0
psycopg2-binary==2.9.5
pydantic==2.5.2
pydantic-settings==2.1.0
pymupdf==1.22.3
python-box==5.1.1
python-docx==1.1.2
python-jose==3.3.0
python-keycloak==4.2.1
requests==2.29.0
setuptools==67.7.2
sqlmodel==0.0.14
tiktoken==0.5.2
unstructured==0.14.10
watchfiles==0.18.1
wheel==0.40.0
xlrd==2.0.1
loguru==0.7.3
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp==1.30.0
black==23.3.0
isort==5.12.0
mypy==1.8.0
pycodestyle==2.11.1
pylint==3.0.3
pylint-pydantic==0.3.2
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-env==1.0.0
types-pyyaml==6.0.12.12
